reorganize code to simplify, consolidate
